split out van_hoeij into different module
